[Buildbucket] Add NextBackendSyncTime to build model
NextBackendSyncTime is a computed field based on the build's UpdateTime.

It'll be used to index builds to query incomplete ones that haven't been updated for a while, so that Buildbucket can sync them with their backend tasks.

NextBackendSyncTime is prefixed with backend target, luci project and and a shard id, to prevent hotspotting when writing to this field and updating its corresponding indexes.

Bug: 1472896
Change-Id: I61016c9cb98faf1ffbeb35e0f8961134b923aea7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4779405
Commit-Queue: Chan Li <chanli@chromium.org>
Reviewed-by: Yuanjun Huang <yuanjunh@google.com>