[cipd] Merge prefix listing and instance listing pages.
The primary goal is to make URLs like
https://chrome-infra-packages.appspot.com/p/infra/tools/cipd/linux-amd64
work as expected (i.e. show package details, not a semi-empty
page).

This is not exactly trivial because of few conflicting requirements:

1. A path like "infra/tools/cipd" can be a prefix (i.e. have
   subpath or subpackages), a package (i.e. have instances) or
   both at the same time.
2. When viewing a prefix, we *must* list its subprefixes to allow
   navigation down the tree.
3. When viewing a package, it is extremely useful to list *sibling*
   packages, since most CIPD package names end with "/<platform>"
   suffix and listing siblings essentially lists supported platforms.

The previous UI worked around this conflict by having two separate
pages: "/p/<path>" will list the children of the prefix, and
"/p/<path>/+/" will show the package and it siblings. And the prefix
listing had two separate items: one for the prefix ("cipd/...") and
another for the package ("cipd").

In the new implementation we somewhat sacrifice clarity of (rare)
case when there's a package and a prefix with the same name:

1. If a path is a prefix (i.e. has children), list children.
2. If a path is a leaf (has no children), list siblings. This
   actually happens only for packages, since CIPDs ListPrefix
   call omits listing "empty" prefixes (since prefixes don't
   really exist in the datastore, only actual packages). When
   listing siblings, show the "current" package in bold. This
   is a cue that it is "selected" in the siblings list.
3. If the path is actually a package (regardless if it is also
   a prefix), show the package info as well.

For most cases this UI looks identical to the previous one. For
cases of a conflict, we show prefix listing (for navigation),
but also instances listing (for exploring the package).

To make this at least slightly less confusing, there's a new UI
icon now that indicates the target of the navigation is a package
and subprefixes can have this icon too, because they can also
be packages! There's also a minor change to breadcrumbs element
when looking at the package, to emphasize what is being looked
at exactly.

One nice bonus is that we can now detect if a path being explored
doesn't exist (or not visible): if path "<parent>/<pkg>" is not
inside `ListPrefix("<parent>")` then it doesn't exist. We get this
for "free" because we have to call `ListPrefix("<parent>")` anyway
to list potential siblings of a package.

Another minor change is moving SVG icons into separate static files,
since we now can theoretically show hundreds of such icons per page,
and having them be inline SVG will bloat the page. Add proper LICENSE
while at it (MIT), since they are copied from Bootstrap Icons.

TEST=launched locally, clicked through all links
R=iannucci@chromium.org
CC=​​​dlf@google.com

Change-Id: I42aa6af0829d919f610158592033d35bbc4412ae
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4813512
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>