[cv] populate TriggeringCL.OriginClid with the biggest CL ID.
This fixes a flaky failure on the following test.
https://source.chromium.org/chromium/infra/infra/+/main:go/src/go.chromium.org/luci/cv/internal/prjmanager/triager/triage_test.go;l=236;drc=fea603e002610df8de9fc3c7171c14d8c7d01b16

It's mainly because `cls` is a map, and the iteration order determines
the value of `OriginClid`.

e.g.,
Given CL-1/2/3, where
- CL-1 is the bottommost CL.
- CL-2 and CL-3 have CQ+2.

If clInfo for CL-2 is processed first then the origin of CL-1 will be
CL-2, whereas the origin CL would be CL-3 if processed first.

Verified the error reproducibility and fix by `go run -count 1000`

Bug: 1470346
Change-Id: I568142d5e051e23184fc089153b9b7819230e336
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4813507
Commit-Queue: Scott Lee <ddoman@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>