[Analysis] Introduce feature flag for policy-based bug management.
The new flag will be used to control rollout of user-visible
aspects of this new functionality. This is the UI change (replacing
the "recommended priority" panel with a "problems" panel) and
the bug management actions.

Once everything is done, we can rip this flag back out.

BUG=b:292843072
TEST=Integration tests
TEST=Spin up frontend and verify window.isPolicyBasedBugManagementEnabled is set.

Change-Id: I887ef21c6ef9bc69f1cb8a3350537a18bfdf1467
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4840763
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>