[swarming] Improve JSON input/output.
ProtoJSONAdapter required to much manual work to build wrappers.
Replace it with a single magical EncodeJSON function that does
all necessary transformations directly based on original types.

Have a dedicated proto file for schemas of input/output (it is
needed now to use these types with EncodeJSON). Unfortunately,
there is still no consistency in formats of inputs/outputs:
  * Some can be represented by proto message.
  * Some are lists of protos, a naked list is not a proto message.
  * Some are dicts of protos, a dict with arbitrary keys is not a
    proto message.

Another remaining complication is summarizeResultsPython that
emits something that has no proto representation whatsoever.
Use a special-purpose hack for it, since it is deprecated.

Yet another complication is that some commands ("bots") sometimes
want to output a non-JSON lists ("-bare" flag). This is now done
through ListWithStdoutProjection with a mental model of "we output
a JSON list to -json-output, but also project some of its data to
stdout as a plain list of strings".

R=jonahhooper@google.com

Change-Id: I390dcece9b09b92e56b9a3d75dcc3555890a78c3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4834133
Reviewed-by: Jonah Hooper <jonahhooper@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>