config: change service_endpoint to hostname
When making a prpc request, all we need is a service hostname instead a
url to the service. This change explicitly request for hostname so
the client won't provide the url by mistake as service_endpoint is kinda
vague on what is expected.

R=yuanjunh

Bug: 1449933
Change-Id: I2120c60b306b6190c3e1c41514ac046dd71becc7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4862959
Reviewed-by: Yuanjun Huang <yuanjunh@google.com>
Auto-Submit: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Yuanjun Huang <yuanjunh@google.com>