[luciexe] expose logdog UI links to luciexe binaries
There's currently no way to easily obtain a logdog link for humans from
a luciexe. Firstly, LogDog streams are not exposed to luciexes except as
io.Writers. Secondly, the luciexe itself only maintains log stream
references as URLs relative to the current build. So two things need to
happen to expose human-usable log links: a richer representation of a
LogDog stream needs to be made available to the luciexe, and it needs to
be possible to resolve the relative URLs to absolute URLs.

For the former, this change modifies the Log method on State and Step to
return a *Log. This *Log is itself an io.Writer, so this mostly
backwards-compatible. Callers that type-assert on this type might break
at compile-time, but this is unlikely to have been the case since
there's nothing useful to type-assert the Writer to. This Log structure
gives us a place to stash richer information about the LogDog stream.

For the latter, this change has the State obtain the current LogDog
namespace from the environment (following the luciexe protocol) and
carry it around. The change then adds a method to Log called UILink that
uses this namespace, in addition to BuildInfra data from the Build
proto, to resolve relative log links.

Change-Id: I3323a87df6acbf0912be0a180303c2590b86f4ae
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4846791
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>