cv: do not add to attention set for passing tryjob
A passing tryjob is often not actionable because the CL is still
awaiting further review, and instead can clutter folks' attention
dashboard.

Stop adding to the attention set in this case. This is a simple
adjustment; there are more refinements that could be made. e.g., if no
one is in the attention set at all, perhaps the owner should be added
back. These further refinements are covered by the bug.

Bug: 1440579
Change-Id: Id9db0dafe697de271bedf174658081bd37cb4238
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4875721
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Michael Pratt <mpratt@google.com>