[rpcexplorer] Implement the core auto-completer logic.
Given a parsed JSON path, we follow it, in each step either
descending into a field or into a list element. While we do
so, we also track where we are in the protobuf descriptor.
In the end we end up knowing what protobuf object matches
the JSON path and can use this information for suggesting
its fields or possible values.

R=randymaldonado@google.com, jonahhooper@google.com

Change-Id: I792832ac5dd85bf5f7638c6f518d41d4b1a55ca6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4709146
Reviewed-by: Randy Maldonado <randymaldonado@google.com>
Reviewed-by: Jonah Hooper <jonahhooper@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>