[luciexe] fix log UI links for steps
Currently log UI links for steps set the namespace for the log to
logPrefix, which includes relLogPrefix. The UI link, however, appends
the namespace to relLogPrefix. This results in relLogPrefix getting
included twice, which is visible in production and produces broken
links.

Separate out logPrefix into relLogPrefix (now called logSuffix) and
logNamespace. The two are then concatenated on-demand to create the new
namespace for steps.

What's slightly odd here is that the tests didn't catch this. The reason
for that is that the tests were reaching into Step.logNamespace to
verify themselves, which led to missing this case. The tests no longer
do that and instead write out the full expected link, which should
prevent mistakes like this in the future.

This actually also reveals a mistake (possibly only in testing) with how
the State-level log links were being handled. At the very least, the way
the log namespace was getting acquired was redundant;
logsink.GetNamespace was already a way to get the right thing. Switch to
that, which fixes the updated State test as well.

Change-Id: I619fcb891659ba815e1c164df4ab710a16bd34d3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4879639
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>