[GCEP] Change expandConfig to only create non-existent VMs
Bug: b/299949621

This patch if applied changes the behavior of expand-config logic

Old Way:
1. Calculate the number of VMs required at the current time
2. Spam a create-vm task for every VM in the config

This ensures that the VM records are in sync with the Config records

New Way:
1. Calculate the number of VMs required at the current time
2. Get all the current VM records for the given config
2. Spam a create-vm task for every VM in the config that currently
   doesn't exist.

This should ensure that the VM records are in sync with the Config
records while only spamming the required tasks. The cost to pay for
this is a read query for the VMs. This will increase the runtime for an
expand-config task. If that queue is bottlenecked, we still have an
option to increase the throughput to 500/s there. It's rate limited to
100/s currently.

On the positive side this should greatly reduce the traffic on that is
currently on create-vm queue. Which means the dynamic scaling can
be more effective.

Change-Id: If5eb458396e5742077ee16033b62f1014954ff44
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4880692
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Reviewed-by: Deniz Kara <denizkara@google.com>