[luci-notify] Fix Builder.GitilesCommit to use gae proto support.
This field (and a few others, but this CL is just an example) use
the old `proto-gae` code generator. Quite a while ago, the gae
datastore library gained the ability to directly handle protos
in a way compatible with the `proto-gae` output.

This has the significant benefit of not stumbling over the
go-vet copylock checker, but also the benefit of no longer actually
unsafely copying around atomic fields.

R=gregorynisbet@google.com, vadimsh

Change-Id: I5d3bb889233fef4cc63c11fb9b26f25839c3eb2a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4888884
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Patrick Meiring <meiring@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>