config: conditionally touch the GS object and retry on pre-condition failure
This is to mitigate the conflict error LUCI Config hits when two request
from the same requester contain the same config file that result in two
requests contend to update the metadata of the same GCS object.

Also, skip touching the object if it has been touched in the past 10
minutes to minize the call to GCS to touch the object.

R=yuanjunh, vadimsh, iannucci

Bug: 1449933
Change-Id: Ibf42a60f7fadd58bf41a82ab837361a9c9ddace2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4892119
Reviewed-by: Yuanjun Huang <yuanjunh@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>