[Bisection] Filter out test failures with old results
I am not sure if we want to expose the end_hour to the RPC to the UI, since it is the time that marks the end of the segments when the analysis started. I think it provides very little value to the users, and may cause confusion, so I don't think we want to show that value.

We can add it later if we want to.

Bug:b/286986518
Change-Id: I1435fd4b1a81e3c74f5d4be346a68177eba072af
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4899457
Commit-Queue: Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Beining Chen <beining@google.com>