[luci-change-verifier] fix a bug in stageTriggerDeps
Initially, the removed code was to prevent the cl from being retriaged,
but it doesn't make sense.

First of all, the decision is not visible to stageTriggerDeps.
stageTriggerDeps() shouldn't modify cls, as it's supposed to provide
input CLs being triaged. Setting the triggeringCL of the staged dep
shouldn't be read as preventing the deps from being retriaged.
The decision is not visible to this function. Also, this function
returns depsToTrigger. If the prevention is necessary, it should be
done by the callsite by parsing the returned depsToTriage.

Secondly, setting triggeringCLs of staged deps won't actually prevent
it from being retriaged. The callsite doesn't use cls to make
the decision.

Third, cls doesn't necessarily contain clInfo for all the deps to
be staged. For example, let's say that CL-A is tracked in the Component.
It doesn't guarantee that all the deps of CL-A are tracked in the
Component, while a logic in triage guarantees that all the deps are tracked
in the PState via notYetLoaded.

As a result, cls[dep.GetClid].triggerinCL causes a panic, if the dep is
not tracked in the component.

Bug: 1470346
Change-Id: I87e25f7939be727fd87b60dc7314a79ae7443931
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4901026
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Scott Lee <ddoman@chromium.org>