[swarming] Handle expired session tokens in UpdateBotSession.
Treat it as an expired session, to signal the bot it needs to
recreated the session (and thus get a fresh session token).

This happens if the bot was stuck for too long (or had no network)
and then suddenly woke up.

Depends on https://chromium-review.googlesource.com/c/4898402 landing
first.

R=jonahhooper@google.com, sfetisov@google.com
BUG=1377118

Change-Id: I9ce9baf4cd8bd9c4e0eb60c19a8b543976408914
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4898932
Reviewed-by: Sergey Fetisov <sfetisov@google.com>
Reviewed-by: Jonah Hooper <jonahhooper@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>