[swarming] Add gRPC-level retries for INTERNAL and UNAVAILABLE.
Based on https://pkg.go.dev/google.golang.org/grpc/examples/features/retry
and stuff found in code search.

Not sure how to test if it works. Need to deploy somewhere and let
it run for a bit to see if the number of failed calls drops.

R=jonahhooper@google.com, iannucci@chromium.org

Change-Id: I5ca52026d1efeb02a8727059a4477f32e508a12b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4913794
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>