[Analysis] Fix links in bugs.
Links posted to bugs were broken.

We have a trailing slash on the UI base URL, when there should
be none. This is in the small section of top-level code that
sets up the bug management that was considered too hard to test,
so there was no test coverage over it.

BUG=b:302604225
TEST=None

Change-Id: I0eb4496e099230d8ec3231d0b578805981140098
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4914593
Auto-Submit: Patrick Meiring <meiring@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Beining Chen <beining@google.com>