cv: add usertext for failed dep run
One of the next CLs will implement a logic to cancel CQ votes
on a child CL if a parent Run fails.

In MCE run, if a parent run fails before CV creates a Run on
a child CL, then the dep triager will retriage the dep of child CL
and retrigger the CQ vote on the parent Run, on which a Run just failed.

To prevent this cycle, one of the next CLs will implement a logic to
cancel the CQ votes on child CLs w/o Runs, when a parent Run fails.
This CL implements usertext for the particular CQ revoke.

Bug: 1470346
Change-Id: Ibe12180e09024d84b3905e8eee1253083778a905
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4902948
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Scott Lee <ddoman@chromium.org>