cv: add an option to override Notify and AddToAttentionSet in clpurger
If a parent Run fails while a child run is waiting for its runs
to be created, OnRunsFinished() will cancel the triggerer on
the child CLs.

In case there are multiple of such CLs in the CL stack, it's not
reasonable to send an email notification for the unvote all the CLs.
Instead, CV should send an email only for the unvote on the originating
CL only.

Find the next CL to find how the new field will be used.

Bug: 1470346
Change-Id: I49e607a47beb09f64d86709285c1698abd336400
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4909985
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Scott Lee <ddoman@chromium.org>