[Analysis] Improve validation for template text.
The template text is long, which makes it difficult to identify
where a problem exists if it is does not match the validation
regex. Move to a rune-level validation, so that we identify the
exact rune which is invalid in the error.

Also loosen validation criteria slightly, to allow new line
characters in HTML.

BUG=b:292843072
TEST=Integration tests

Change-Id: I9434f61e290ee36a06465cc5a250da1a632b0d3c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4918321
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>