cv: stop triaging deps that have triaged by the previous manage-project
The current stageTriggerDeps assumes that
- triageDeps() would report a given dep as notYetLoaded, if the dep was
not monitored by the component.
- therefore, param `cls`, should contain clinfo for all the deps.

However, it is incorrect. triageDeps() only checks if a given dep is
tracked in the PState. It doesn't care whether the dep is monitored
by the component.

Therefore, stageTriggerDeps caused duplicate TriggeringCL tasks for
the same dep via multiple rounds of manage-project runs, and this
CL fixes the bug.

Bug: 1470346
Change-Id: I96430e385630020a54d6db2ef9dd061b57bf9416
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4923509
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Scott Lee <ddoman@chromium.org>