[Analysis] Add invocation ID to test verdict results
This is for b/301523224. We want to exclude tests that run in the chromium.tests.gpu pool due to resource constrain (note that this is the pool that the swarming test task runs, not the pool that builder/tester run).

The pool information does not seems to be visible in either resultDB or LUCI Analysis and is buried in the swarming table https://screenshot.googleplex.com/3Gqr5ki4QmfyzK9. I think we need the parent invocation ID information so we can query swarming.


Bug:b/301523224
Change-Id: I235aa720e45242e27a3f6e9856a9d29d04ed980c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4938050
Reviewed-by: Patrick Meiring <meiring@google.com>
Commit-Queue: Tuan Nguyen <nqmtuan@google.com>