config: do not save attempt entity if config is up to date
I discovered a bug in LUCI Config that if a config is imported with
warning, the subsequent import make the import green because it
clears the validation result and override the message with up-to-date.

I think it is okay to not saving the attempt entity in that case.
Keeping the old attempt entity should be good enough.

R=yuanjunh

Change-Id: I5f5b79c1a8a907753f39bed804d97bc91ccbd996
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4960499
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Yuanjun Huang <yuanjunh@google.com>