[Analysis] Add ability to filter recent failures to focus on metric.
Help people drill down into problems by allowing the recent failures to be filtered down to those which contribute to a metric of interest.
The filter and groupby and sort order are exposed as URL parameters
to also allow deep linking to the view from bug comments.

A "Show failures" button is added next to a problem to invite the user
to drill down into the failures for a problem.

The recent failures groups are also enhanced with a badge to
make it clearer what each group is.

BUG=b:292843072
TEST=npm run test

Change-Id: I3f6f72ffe69ff62ac5acdf8f29940df6ab4bbe43
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4970375
Reviewed-by: Matthew Warton <mwarton@google.com>
Auto-Submit: Patrick Meiring <meiring@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>