[gae] Dedup code for "exploding" and "collecting" key from meta.
It will also be reused in a bunch of places that deal with nested
entities.

This code essentially already existed in slightly different forms.

"Exploding" a key is done via PopulateKey(...). This CL changes it
to be more assertive and try to populate all possible meta fields,
not just `$key`. This matters when the receiver is a PropertyMap
which is then later used to populate some other MGS (which might
not have `$key`, but has `$kind` and `$id`).

"Collecting" a key was already essentially done by KeyForObjErr,
but it had no public variant that can use an existing KeyContext.

R=iannucci@chromium.org

Change-Id: I17ec3548b8f46a7bd02589cf910b687b50cff82e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4982006
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>