[cv] Reapply debit op with every poke
This change reapplies debit quota op for every poke event and before
credit op. In the event where redis is wiped out and the debit op is lost, this ensures that it is restored. debit op is idempotent and hence would be deduped if the op has already been performed.

Bug: 1466346
Change-Id: Ica2d900b9852dab54cf02b4371c2beb5ce0ad2fb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4977576
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Aravind Vasudevan <aravindvasudev@google.com>