[gae] Introduce optional and nullable properties.
There are several properties in Swarming datastore model which
give different semantic meaning to an absence of property and
its zero value.

E.g. missing try_number means the task was never attempted to run,
but a try_number set to 0 means the task executed its first attempt.
There's a query that filters on `try_number == 0`. This property
can now be declared either as Optional[int] or Nullable[int]
(the difference is whether the "absense" is indexed, which in
this case doesn't matter).

As another example, there's a query on time.Time `completed_ts`
field that finds all entities where `completes_ts == null`. This
property can now be clared as Nullable[time.Time], so that its
absence can be filtered on in queries.

This was not possible to express in Go using standard
PropertyLoadSaver.

R=iannucci@chromium.org, jonahhooper@google.com

Change-Id: Ieb2bc0f367a106ebd20c6eb53f06e3c5494a8545
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5005225
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Jonah Hooper <jonahhooper@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>