[cv] Fix pending_msg to be posted only once
The current implementation for posting pending msg when quota is not
available has two issues:
1. Gerrit prepends metadata to the posted msg which breaks
   post_gerrit_messsage's deduplication logic of using equality check.
2. Longops are executed before start handler in run manager which
   removes the pending msg entry from run.OngoingLongOps before start
   handler is executed, leading to start handler making duplicate
   enqueues.

This CL fixes both of the aforementioned issues by updating equality check to contains check within post_gerrit_messsage and adding a new PendingMsgPosted entry to runState which restricts enqueuing pending msg to only once.

The downside of this change is that, if posting pending msg longop fails, we do not re-enqueue another longop. However, this should be an
edge case given we retry posting the longop anyway.

Bug: 1466346
Change-Id: I55d800346b5369184cd161959b0d0cf029e701d3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5007182
Reviewed-by: Scott Lee <ddoman@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Aravind Vasudevan <aravindvasudev@google.com>