[Analysis] Stop auto-verifying non-LUCI Analysis bugs.
The auto-verify and auto-prioritise features make sense for LUCI
Analysis auto-filed bugs, but not for manually created bugs that
we do not know the resolution criteria of.

For manually created failure association rules, we already default
the update bug toggle to OFF, but we do not do so when a LUCI
Analysis bug is merged into another bug, even though this is
a similiar case. (Rule for non-LUCI Analysis created bug.)

See b/303834216 and b/300560193 for two examples today where
this was behaving undesirably.

BUG=None
TEST=Integration tests

Change-Id: Ia6f711ac44009e39797cda2224f0de2763283ec6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5009655
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>