[Analysis] Remove unused fields from GetConfig RPC.
These fields were used to support the recommended priority
section, which has been replaced by the problems section.

We need to stagger the rollout of this change as clients with
cached versions of the frontend SPA will still be depending on
this data until they get the latest frontend code.

BUG=b:292843072
TEST=Integration tests

Change-Id: I9b12c48d8897a35769df30a20f7361f6a80a1144
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5004480
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>