[swarming] Handle lost RBE leases.
Sometimes RBE thinks "the bot running the task appears to be lost"
and it treats it by silently ignoring the ACTIVE lease as reported
by the bot.

This was resulting in HTTP 500 errors in Swarming wrapper. Now this
condition is propagated to the bot, which will handle it more
gracefully.

R=jonahhooper@google.com

Change-Id: I0efa1e466654dd38b6b6ab4ce08f971968fceae8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5021313
Reviewed-by: Jonah Hooper <jonahhooper@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>