[cv] cltriggerer and gerrit/trigger mark the CL snapshot as outdated
* cltriggerer
will mark the CL snapshot as outdated on vote success.
will schedule UpdateCL task on vote success.

* gerrit/trigger
will mark the CL snapshot as outdated on vote success.

* triager/stagePurges()
stagePurges() schedules PurgingCL() if a given info has purgeReasons
set. However, as of this moment, the reason field is set only either
- the CL has invalid deps
- run triager finds that the Run to create already exists, and
  the existing run ended already.

Since evalpcl sets deps with nil, if the CL snapshot is outdated,
none of the conditions will be true, and this change is supposed
to be noop. Nevertheless, this change is just for sake, and for
future changes.

* stageTriggerCLDeps
If a given CL is outdated, it wouldn't have any deps to triager,
and pcl.deps is nil. However, this change is to check if any deps
have outdated snapshot.

Note OnTriggeringCLDepsCompleted() will not remove the op until
all the deps are fresh. Therefore, this change is expected to be
noop, just like the changes for stagePurges(). However, this change
is just for sake and future changes.

Bug: 1470346
Change-Id: Iefa305d4b298225caab958eb217f032b3d0f2418
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5021835
Commit-Queue: Scott Lee <ddoman@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>