[swarming] Implement config datastore caching logic.
Configs are fetched from LUCI Config via a cron job and are
stored in the datastore, along with their "digest" (a hash of
content).

Each backend process loads the initial copy of the configs from
the datastore and periodically looks in the datastore at the most
recently stored digest. If it is different, it refetches full
configs from the datastore.

Few complications:
* Need a digest because revision can change even when configs are
  unchanged. We can theoretically use revision as well and just
  do some unnecessary reloads, but adding a digest isn't hard.
* Need to handle cases related to new service bootstrap, when
  there's no config yet.

R=jonahhooper@google.com, sfetisov@google.com

Change-Id: Ib7e3ab388425fa5631c0aeefbfaf7c3543729379
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5034561
Reviewed-by: Sergey Fetisov <sfetisov@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>