[server/quota] Add WITH_POLICY_LIMIT_DELTA option for ApplyOps
This change introduces WITH_POLICY_LIMIT_DELTA option for Operations.
This option, when enabled, will apply the delta of a policy change to
the account's balance.

For example, say account Foo's policy changes from limit 5 to 10 and its
balance is 3. When WITH_POLICY_LIMIT_DELTA option is set, Foo's balance
is adjusted to 8 (3 + delta, where delta = 10 - 5). Similarly, if the
limit is updated from 5 to 1, Foo's balance will be adjusted to -1 (3 +
delta, where delta = 1 - 5).

This is useful for cases such as CV where we don't rely on refill
strategies to recover balance. When an account's policy changes, CV will
simply trigger server/quota with the new policy and with this option,
server/quota will handle adjusting the difference.

Bug: 1466346
Change-Id: I1a8e653f79fa3cf111087483d24202129e6eacb0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5027943
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Aravind Vasudevan <aravindvasudev@google.com>