[Analysis] Add methods for reading reclustering progress in the past.
To resolve update-analysis-and-bugs availability / latency alerts,
it is proposed to break the task into two: one task will do the
analysis (update cluster summaries table) and one will do the
bug updates.

The bug updates task will need the reclustering progress as it was
when the analysis task started, as cluster_summaries will be
reflecting the clustered_failures table as it was then, not as it
is now. This is in turn needed so that job knows which
bugs/clusters have invalid impact data due to ongoing reclustering,
so priority updates/auto-closure for these can be suspended.

This CL lays the groundwork for fetching past reclustering progress.

Alternatives considered:
- run the current query in the analysis job and pass the reclustering progress to the second job. This was seen as too complex; we need to pass progress for all projects and the progress consists of six fields.
- use Spanner stale reads to achieve reads of historic reclustering progress. This was discounted due to concerns about testability of stale reads in unit/integration tests where test data is populated only milliseconds earlier.

BUG=None
TEST=Integration tests

Change-Id: I69eb2d76d9bdbcae2c71f020fbcdefa9650b763b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5056101
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>