[common] Add ability to synchronise BQ table description, labels.
In the process of preparing these updates, it was identified that
the current way in which views are always being updated is not
necessarily safe, even if all current users are using it safely.

After this change, callers must explicitly request synchronisation
of all non-schema settings as those updates may not converge in the
case of canary + stable deployments with different table specs.

It is suggested that such synchronization only occur from a
periodic cron job instead of from task queue (whose backend may be
sharded over many GAE or GKE instances) to minimise consequences
of an edit war if canary + stable have different table specs.

The option can not be provided to the EnsureTable method on
SchemaApplyer as this cache is designed for use with task queue-driven exports and is not necessary for cron job-driven updates.

BUG=None
TEST=Integration tests

Change-Id: I9f4be4dcfbc47fe8fc8e1201c3e64290e4be2aa5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5064817
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Tuan Nguyen <nqmtuan@google.com>