[analysis] Fix use of sources proto unaligned with ResultDB.
LUCI Analysis is using a sources message which not aligned
with that of ResultDB for the capture of gerrit changelists.
This results in slight data loss and will result in friction
when clients need to query LUCI Analysis for exonerations at
a source position.

Update the protos to standardise on the same format while we can.

The LUCI Analysis proto can always still extend the message to
add the owner_kind information later if it is needed, but it is
presently not being populated.

The existing Changelist message is still required for the
legacy Test History service and exoneration RPC, which obtain
changelist data from a different source.

BUG=None
TEST=Integration tests

Change-Id: Ib27e899a10e92eba2685636ba3f89f0be1ed11f5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5076013
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>